flask_server/requirements.txt
bcrypt==3.2.0
bidict==0.21.2
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
python-dotenv==0.17.1
python-engineio==4.1.0
python-socketio==5.2.1
six==1.15.0
Werkzeug==1.0.1
